Time: 30 ms (59.12%), Space: 44.4 MB (22.33%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -3,7 +3,7 @@ public:
     int numDistinct(string text1, string text2) {
         int n= text1.size();
         int m = text2.size();
-        vector<vector<int>>dp(n+1,vector<int>(m+1,0));
+        vector<vector<double>>dp(n+1,vector<double>(m+1,0));
         for(int i=0;i<=n;i++) dp[i][m]=1;
         for(int i=n-1;i>=0;i--){
             for(int j=m-1;j>=0;j--){
